requirements.txt
appdirs==1.4.3
attrs==19.1.0
black==19.3b0
Click==7.0
pycryptodome==3.8.2
toml==0.10.0
